refactor: Компонент OfferCard перенесен в Main
Кажется, пока что здесь можно не плодить лишние файлы.
