--- src/components/offer-card/offer-card.jsx
@@ -1,57 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const OfferCard = ({offer}) => (
-  <article className="cities__place-card place-card">
-    {offer.isPremium &&
-      <div className="place-card__mark">
-        <span>Premium</span>
-      </div>
-    }
-    <div className="cities__image-wrapper place-card__image-wrapper">
-      <a href="#">
-        <img className="place-card__image" src={offer.img} width={260} height={200} alt="Place image" />
-      </a>
-    </div>
-    <div className="place-card__info">
-      <div className="place-card__price-wrapper">
-        <div className="place-card__price">
-          <b className="place-card__price-value">€{offer.price}</b>
-          <span className="place-card__price-text">/&nbsp;night</span>
-        </div>
-        <button className={`place-card__bookmark-button button${offer.isFavorite ? ` place-card__bookmark-button--active` : ``}`} type="button">
-          <svg className="place-card__bookmark-icon" width={18} height={19}>
-            <use xlinkHref="#icon-bookmark" />
-          </svg>
-          <span className="visually-hidden">{offer.isFavorite ? `In bookmarks` : `To bookmarks`}</span>
-        </button>
-      </div>
-      <div className="place-card__rating rating">
-        <div className="place-card__stars rating__stars">
-          <span style={{width: offer.rating * 20 + `%`}} />
-          <span className="visually-hidden">Rating</span>
-        </div>
-      </div>
-      <h2 className="place-card__name">
-        <a href="#">{offer.title}</a>
-      </h2>
-      <p className="place-card__type">{offer.type}</p>
-    </div>
-  </article>
-);
-
-OfferCard.propTypes = {
-  offer: PropTypes.shape({
-    id: PropTypes.number.isRequired,
-    title: PropTypes.string.isRequired,
-    city: PropTypes.string.isRequired,
-    type: PropTypes.string.isRequired,
-    price: PropTypes.number.isRequired,
-    rating: PropTypes.number.isRequired,
-    img: PropTypes.string.isRequired,
-    isPremium: PropTypes.bool,
-    isFavorite: PropTypes.bool
-  })
-};
-
-export default OfferCard;
